NeighborlyFrontEnd/requirements.txt
azure-functions
azure-eventhub
certifi
chardet
click
dominate
Flask
Flask-Bootstrap
feedgen
feedwerk
idna
itsdangerous
Jinja2
MarkupSafe
requests
urllib3
visitor
Werkzeug==2.2.0
